requirements.txt
oracledb
pandas
matplotlib
seaborn
plotly
python-dotenv
sqlalchemy
